+++ SlipNFrag-Win64/vid_win64.cpp
@@ -70,7 +70,7 @@ void VID_Init(unsigned char* palette)
 	d_pzbuffer = zbuffer.data();
 	int surfcachesize = D_SurfaceCacheForRes(vid_width, vid_height);
 	surfcache.resize(surfcachesize);
-	D_InitCaches(surfcache.data(), (int)surfcache.size());
+	D_InitCaches(surfcache.data(), surfcache.size());
 }
 
 void VID_Resize()
@@ -92,7 +92,7 @@ void VID_Resize()
 	int surfcachesize = D_SurfaceCacheForRes(vid_width, vid_height);
 	surfcache.resize(surfcachesize);
 	Draw_ResizeScanTables();
-	D_InitCaches(surfcache.data(), (int)surfcache.size());
+	D_InitCaches(surfcache.data(), surfcache.size());
 	R_ResizeTurb();
 	R_ResizeEdges();
 	vid.recalc_refdef = 1;
@@ -111,7 +111,7 @@ void VID_ReallocSurfCache()
 	}
 	surfcache.resize(surfcache.size() + increase);
 	Draw_ResizeScanTables();
-	D_InitCaches(surfcache.data(), (int)surfcache.size());
+	D_InitCaches(surfcache.data(), surfcache.size());
 }
 
 void VID_Shutdown(void)
